+++ src/main/java/com/team/cklob/gami/domain/auth/controller/AuthController.java
@@ -1,7 +1,7 @@
 package com.team.cklob.gami.domain.auth.controller;
 
 import com.team.cklob.gami.domain.auth.dto.request.*;
-import com.team.cklob.gami.domain.auth.dto.responnse.TokenResponse;
+import com.team.cklob.gami.domain.auth.dto.response.TokenResponse;
 import com.team.cklob.gami.domain.auth.service.*;
 import com.team.cklob.gami.domain.auth.exception.TooManyRequestsException;
 import jakarta.servlet.http.HttpServletRequest;
