+++ src/main/java/com/team/cklob/gami/domain/auth/service/ReissueTokenService.java
@@ -1,6 +1,6 @@
 package com.team.cklob.gami.domain.auth.service;
 
-import com.team.cklob.gami.domain.auth.dto.responnse.TokenResponse;
+import com.team.cklob.gami.domain.auth.dto.response.TokenResponse;
 
 public interface ReissueTokenService {
     TokenResponse execute(String refreshToken);
