+++ src/main/java/com/team/cklob/gami/domain/auth/service/impl/ReissueTokenServiceImpl.java
@@ -3,7 +3,7 @@ package com.team.cklob.gami.domain.auth.service.impl;
 import com.team.cklob.gami.domain.auth.entity.RefreshToken;
 import com.team.cklob.gami.domain.auth.exception.NotFoundUserException;
 import com.team.cklob.gami.domain.auth.exception.UnauthorizedException;
-import com.team.cklob.gami.domain.auth.dto.responnse.TokenResponse;
+import com.team.cklob.gami.domain.auth.dto.response.TokenResponse;
 import com.team.cklob.gami.domain.auth.repository.RefreshTokenRepository;
 import com.team.cklob.gami.domain.auth.service.ReissueTokenService;
 import com.team.cklob.gami.domain.member.entity.Member;
